Refactor: organize proyect into several files and directories
