+++ src/database.py
@@ -1,0 +1,31 @@
+import mysql.connector
+from mysql.connector import Error
+from sqlalchemy import create_engine
+from sqlalchemy.ext.declarative import declarative_base
+from sqlalchemy.orm import sessionmaker
+
+class settings():
+    database_username = "root"
+    database_password = "a1s2d3fr"
+    database_hostname = "localhost"
+    database_name = "scp_final"
+    database_port = 3306
+    
+    
+
+SQLALCHEMY_DATABASE_URL = f"mysql+mysqlconnector://{settings.database_username}:{settings.database_password}" \
+                          f"@{settings.database_hostname}:{settings.database_port}/{settings.database_name}"
+
+engine = create_engine(SQLALCHEMY_DATABASE_URL)
+
+SessionLocal = sessionmaker(autocommit=False, autoflush=False, bind=engine)
+
+Base = declarative_base()
+
+# Dependency
+def get_db():
+    db = SessionLocal()
+    try:
+        yield db
+    finally:
+        db.close()
