+++ src/routes/home.py
@@ -1,0 +1,11 @@
+from fastapi import APIRouter, Depends
+from sqlalchemy.orm import Session
+from ..database import get_db
+from .. import models
+
+app = APIRouter()
+
+@app.get("/")
+def home(db: Session=Depends(get_db)):
+    test = db.query(models.Test).all()
+    return test
